==> NMT/1_Data_creation___NMT.ipynb <==
Download & Import libraries

import numpy as np
import pandas as pd
import nltk
nltk.download('punkt')
from nltk.tokenize import sent_tokenize

Read files & store in dataframe

bible_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/bible-uedin.en-mr.mr')
bible_eng_f =open('/content/drive/MyDrive/NLP/NMT/Marathi-English/bible-uedin.en-mr.en')

gnome_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/GNOME.en-mr.mr')
gnome_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/GNOME.en-mr.en')

kde_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/KDE4.en-mr.mr')
kde_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/KDE4.en-mr.en')

qed_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/QED.en-mr.mr')
qed_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/QED.en-mr.en')

ubuntu_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/Ubuntu.en-mr.mr')
ubuntu_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/Ubuntu.en-mr.en')

wiki_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/wikimedia.en-mr.mr')
wiki_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/wikimedia.en-mr.en')

cc_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/MultiCCAligned.en-mr.mr')
cc_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/MultiCCAligned.en-mr.en')

ted_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/TED2020.en-mr.en')
ted_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/TED2020.en-mr.mr')

tico_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/tico-19.en-mr.en')
tico_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/tico-19.en-mr.mr')

wikimatrix_eng_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/WikiMatrix.en-mr.en')
wikimatrix_mar_f = open('/content/drive/MyDrive/NLP/NMT/Marathi-English/WikiMatrix.en-mr.mr')

lines = bible_mar_f.readlines()
bible_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = bible_eng_f.readlines()
bible_eng = pd.DataFrame(lines, columns = ["English"])

lines = gnome_mar_f.readlines()
gnome_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = gnome_eng_f.readlines()
gnome_eng = pd.DataFrame(lines, columns = ["English"])

lines = kde_mar_f.readlines()
kde_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = kde_eng_f.readlines()
kde_eng = pd.DataFrame(lines, columns = ["English"])

lines = qed_mar_f.readlines()
qed_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = qed_eng_f.readlines()
qed_eng = pd.DataFrame(lines, columns = ["English"])

lines = ubuntu_mar_f.readlines()
ubuntu_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = ubuntu_eng_f.readlines()
ubuntu_eng = pd.DataFrame(lines, columns = ["English"])

lines = wiki_mar_f.readlines()
wiki_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = wiki_eng_f.readlines()
wiki_eng = pd.DataFrame(lines, columns = ["English"])

many_things = pd.read_csv('/content/drive/MyDrive/NLP/NMT/Marathi-English/mar(many_things).txt',sep="\t" ,names = ["English","Marathi","extra"])

lines = cc_mar_f.readlines()
cc_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = cc_eng_f.readlines()
cc_eng = pd.DataFrame(lines, columns = ["English"])

lines = ted_mar_f.readlines()
ted_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = ted_eng_f.readlines()
ted_eng = pd.DataFrame(lines, columns = ["English"])

lines = tico_mar_f.readlines()
tico_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = tico_eng_f.readlines()
tico_eng = pd.DataFrame(lines, columns = ["English"])

lines = wikimatrix_mar_f.readlines()
wikimatrix_mar = pd.DataFrame(lines, columns = ["Marathi"])
lines = wikimatrix_eng_f.readlines()
wikimatrix_eng = pd.DataFrame(lines, columns = ["English"])

Concatenate english and marathi dataset

1)Many things

many_things_new=many_things.drop(["extra"], axis=1)

many_things_new.tail(10)

many_things_new = many_things_new.replace('\n','', regex=True)

many_things_new

2)Bibble

bible = pd.concat([bible_mar, bible_eng], axis=1)

bible.tail(10)

bible.isna().sum()

bible = bible.replace('\n','', regex=True)

bible.tail(10)

3)Gnome

gnome = pd.concat([gnome_mar, gnome_eng], axis=1)

gnome.tail(10)

gnome.isna().sum()

gnome = gnome.replace('\n','', regex=True)

gnome = gnome.replace('%','', regex=True)

gnome.tail(10)

4)Kde

kde = pd.concat([kde_mar, kde_eng], axis=1)

kde.head(10)

kde.isna().sum()

kde = kde.replace('\n','', regex=True)

kde = kde.replace('%','', regex=True)

kde.tail(10)

5)Qed

qed = pd.concat([qed_mar, qed_eng], axis=1)

qed.head(10)

qed.isna().sum()

qed = qed.replace('\n','', regex=True)

qed["English"] = qed["English"].str.lower()

qed.tail(10)

6)Ubuntu

ubuntu = pd.concat([ubuntu_mar, ubuntu_eng], axis=1)

ubuntu.tail(100)

ubuntu.isna().sum()

ubuntu = ubuntu.replace('\n','', regex=True)

ubuntu.tail(10)

7)Wiki

wiki = pd.concat([wiki_mar, wiki_eng], axis=1)

wiki.tail(10)

wiki.isna().sum()

wiki = wiki.replace('\n','', regex=True)

wiki.tail(10)

8)CC

cc = pd.concat([cc_mar, cc_eng], axis=1)

cc.tail(100)

cc.isna().sum()

cc = cc.replace('\n','', regex=True)

cc.tail(10)

9)TED

ted = pd.concat([ted_mar, ted_eng], axis=1)

ted.tail(100)

ted.isna().sum()

ted = ted.replace('\n','', regex=True)

ted.tail(100)

10)TICO

tico = pd.concat([tico_mar, tico_eng], axis=1)

tico.tail(100)

tico.isna().sum()

tico = tico.replace('\n','', regex=True)

tico.tail(100)

11)Wiki Matrix

wikimatrix = pd.concat([wikimatrix_mar, wikimatrix_eng], axis=1)

wikimatrix.tail(100)

wikimatrix.isna().sum()

wikimatrix = wikimatrix.replace('\n','', regex=True)

wikimatrix.tail(100)

sentences: 1019506

df = wikimatrix.append([many_things_new, bible, gnome, kde, qed, ubuntu, wiki, cc, ted, tico])

df.info

Status

**`wiki,qed,kde,bible`** -->done <br>
**`ubuntu,gnome`** -->not done

Clean Ubuntu & Gnome dataset

Making model based on QED for sample

Saving Dataset

qed.to_csv(r"/content/drive/My Drive/Marathi-English/Cleaned Dataset/qed_cleaned.csv")
