==> NMT/2_Split_Text___NMT.ipynb <==
from pickle import load
from pickle import dump
from numpy.random import rand
from numpy.random import shuffle  
import pandas as pd

raw_dataset =   pd.read_csv('/content/drive/My Drive/Marathi-English/Cleaned Dataset/qed_cleaned.csv')

raw_dataset=raw_dataset.drop(['Unnamed: 0'], axis = 1)

raw_dataset.tail(10)

train, test = raw_dataset[:1300], raw_dataset[1300:1745]

train.tail(10)

test.tail(10)

dump(raw_dataset, open('/content/drive/MyDrive/Marathi-English/Cleaned Dataset/qed_raw_dataset.pkl', 'wb'))
dump(train, open('/content/drive/MyDrive/Marathi-English/Cleaned Dataset/qed_train.pkl', 'wb'))
dump(test, open('/content/drive/MyDrive/Marathi-English/Cleaned Dataset/qed_test.pkl', 'wb'))
