==> NMT/4_Evaluate Neural Translation Model_NMT.ipynb <==
!free -h

from google.colab import drive
drive.mount('/content/drive')

from pickle import load
from numpy import array
from numpy import argmax
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.models import load_model
from nltk.translate.bleu_score import corpus_bleu

# load a clean dataset

def load_clean_sentences(filename):
	return load(open(filename, 'rb'))

# fit a tokenizer

def create_tokenizer(lines):
	tokenizer = Tokenizer()
	tokenizer.fit_on_texts(lines)
	return tokenizer

# max sentence length

def max_length(lines):
	return max(len(line.split()) for line in lines)

# encode and pad sequences

def encode_sequences(tokenizer, length, lines):
	# integer encode sequences
	X = tokenizer.texts_to_sequences(lines)
	# pad sequences with 0 values
	X = pad_sequences(X, maxlen=length, padding='post')
	return X

# map an integer to a word

def word_for_id(integer, tokenizer):
	for word, index in tokenizer.word_index.items():
		if index == integer:
			return word
	return None

# generate target given source sequence

def predict_sequence(model, tokenizer, source):
	prediction = model.predict(source, verbose=0)[0]
	integers = [argmax(vector) for vector in prediction]
	target = list()
	for i in integers:
		word = word_for_id(i, tokenizer)
		if word is None:
			break
		target.append(word)
	return ' '.join(target)

# evaluate the skill of the model
def evaluate_model(model, tokenizer, sources, raw_dataset):
  actual, predicted = list(), list()
  for i, source in enumerate(sources):
    # translate encoded source text
    print("Source:")
    print(source)
    source = source.reshape((1, source.shape[0]))
    translation = predict_sequence(model, eng_tokenizer, source)
    print("Translation")
    print(translation)
    raw_src, raw_target = raw_dataset.iloc[i]
    if i < 10:
      print('src=[%s], target=[%s], predicted=[%s]' % (raw_src, raw_target, translation))
    actual.append([raw_target.split()])
    print(translation)
    # print("--------------------------------")
    predicted.append(translation.split())
    print("Predicted")
    print(predicted)
  # calculate BLEU score
  print('BLEU-1: %f' % corpus_bleu(actual, predicted, weights=(1.0, 0, 0, 0)))
  print('BLEU-2: %f' % corpus_bleu(actual, predicted, weights=(0.5, 0.5, 0, 0)))
  print('BLEU-3: %f' % corpus_bleu(actual, predicted, weights=(0.3, 0.3, 0.3, 0)))
  print('BLEU-4: %f' % corpus_bleu(actual, predicted, weights=(0.25, 0.25, 0.25, 0.25)))
  # print(predicted)
  # print("___")
  # print(actual)


# load datasets
dataset = load_clean_sentences('/content/drive/MyDrive/NLP/NMT/Marathi-English/Cleaned Dataset/qed_raw_dataset.pkl')
train = load_clean_sentences('/content/drive/MyDrive/NLP/NMT/Marathi-English/Cleaned Dataset/qed_train.pkl')
test = load_clean_sentences('/content/drive/MyDrive/NLP/NMT/Marathi-English/Cleaned Dataset/qed_test.pkl')

# prepare marathi tokenizer

mar_tokenizer = create_tokenizer(dataset["Marathi"])
mar_vocab_size = len(mar_tokenizer.word_index) + 1
mar_length = max_length(dataset["Marathi"])
print('Marathi Vocabulary Size: %d' % mar_vocab_size)
print('Marathi Max Length: %d' % (mar_length))

# prepare english tokenizer

eng_tokenizer = create_tokenizer(dataset["English"])
eng_vocab_size = len(eng_tokenizer.word_index) + 1
eng_length = max_length(dataset["English"])
print('English Vocabulary Size: %d' % eng_vocab_size)
print('English Max Length: %d' % (eng_length))

train["Marathi"]

test["Marathi"]

#Prepare data

trainX = encode_sequences(mar_tokenizer, mar_length, train["Marathi"])
testX = encode_sequences(mar_tokenizer, mar_length, test["Marathi"])

# load model
model = load_model('/content/drive/MyDrive/NLP/NMT/Marathi-English/model/qed/model.h5')

# test on some training sequences
print(train)

evaluate_model(model, eng_tokenizer, trainX, train)

print(test)

evaluate_model(model, eng_tokenizer, testX, test)
